continue work on notifs, tidy some stuff up elsewhere
