+++ testrig/federatingdb.go
@@ -2,10 +2,10 @@ package testrig
 
 import (
 	"github.com/superseriousbusiness/gotosocial/internal/db"
-	"github.com/superseriousbusiness/gotosocial/internal/federation"
+	"github.com/superseriousbusiness/gotosocial/internal/federation/federatingdb"
 )
 
 // NewTestFederatingDB returns a federating DB with the underlying db
-func NewTestFederatingDB(db db.DB) federation.FederatingDB {
-	return federation.NewFederatingDB(db, NewTestConfig(), NewTestLog())
+func NewTestFederatingDB(db db.DB) federatingdb.DB {
+	return federatingdb.New(db, NewTestConfig(), NewTestLog())
 }
